+++ Timeline/ClientApp/src/app/user/internal-user-service/http-entities.ts
@@ -1,5 +1,8 @@
 import { UserCredentials, UserInfo } from '../entities';
 
+export const createTokenUrl = '/api/User/CreateToken';
+export const validateTokenUrl = '/api/User/ValidateToken';
+
 export type CreateTokenRequest = UserCredentials;
 
 export interface CreateTokenResponse {
